requirements.txt
pywa==2.11.0
fastapi
python-dotenv
uvicorn
